moving getConditionsWithRequirements function
